% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{packageResults}
\alias{packageResults}
\title{Package the results for sharing with OHDSI researchers}
\usage{
packageResults(results, table1 = NULL, saveFolder, dbName, modelComp = T)
}
\arguments{
\item{results}{The result of running applyExistingstrokeModels()}

\item{table1}{The result of running getTable1()}

\item{saveFolder}{Name of the folder where results are saved to}

\item{dbName}{Name of the databased used to validate the models}

\item{modelComp}{whether to include comparisons between the model performances}
}
\description{
Package the results for sharing with OHDSI researchers
}
\details{
This function packages the results.
}
